Removing all NCIT eq and adding them to external (importable) file
To move forward with organising the HP imports properly, we now removed all logical definitions with NCIT references in it from HP, and added them, entirely unchanged, to a new file called hp_ncit_eq.owl. The file can, if necessary, be easily imported into HP - for now they are seperate until the neoplasm discussion is resolved. There is zero loss of information.